Rework zoom callbacks, and integrate dblclick-reset
This makes it much easier for users to add mouse zooming, because they
don't have to write any callbacks. Formerly it was awkward because
the user had to test which kind of button-click it is in the callback,
and it risked tromping on any pre-existing mouse button handler.

The zoom callbacks were reworked to use `cfunction`-style callbacks,
largely because the backtraces are much more useful when there is
an error.