panzoom: use mouse callback stack, 2 functions instead of 4
The main motivation for this change is to reduce the effect of
callback order on behavior. It uses the new Gtk MouseHandler
stack to temporarily assume control of mouse operations, rather than
defining low-level callback functions. In tests, this already
seems more robust, and allow one to swap between user-defined
callbacks and the zoom/pan callbacks very easily.