Prevent Immerse toolbar from stealing the focus on up-arrow
The trick is to signal that the event has been handled. Despite the
fact that most/all of the GUI elements have "can-focus" set false,
hitting the up arrow previously caused the final button in the toolbar
to be selected. This meant that up-arrow caused at most one pan event,
and after that all input got directed to the button.

If the handler is blocked, then the toolbar can still steal focus,
but now clicking on the canvas restores the focus to the canvas.