WIP: A first stab at managing state via Reactive.
I began with panzoom functionality.  The idea is that panzoom will associate a Reactive Signal with a given canvas and store it in GtkUtilities.guidata.  The signal just contains the currently computed region of interest (ROI), and other GUI elements can hook into that signal to do other things, like, for example, redrawing an image.

Currently only the `panzoom_key` function is working, but `panzoom_mouse` should be easy now if we decide to take this design direction.  Some other decisions I made:

-Combine `xview` and `yview` into a new ViewROI type.  I thought at first that this would be a very helpful change, but in the end I only slightly prefer it to keeping x and y in separate signals.  guidata now stores two roi variables per canvas:  cur_roi and max_roi.
-guidata's :cur_roi is a Reactive.Signal with a ViewROI value.  :max_roi is not a Signal, but we may want to change that at some point
-The GTK keypress callback now updates dedicated signals listening for each keypress.  Currently I'm not using the values of these signals for anything; they simply cause an ROI update whenever a "true" value is pushed to them.  At some point maybe this behavior will be changed so that the signal values accurately represent the state of the key. (is it pressed right now or not?)
-I implemented the keypress listener signals as a wrapped Reactive.Signal with extra information about which key and which canvas it's associated with.  To me this seems useful for debugging purposes.  If we decide we like that style, then we can probably just go ahead and subtype Reactive.Signal.

I updated one of the demo scripts to demonstrate that they panzoom_key works.  Note that I switched around some of the default control keys because on my laptop SHIFT and CONTROL combinations don't work.  I have the same problem on the current master, so this is not related to the Reactive migration.